+++ src/element.hpp
@@ -6,12 +6,12 @@
 class Element : public std::vector<Triangle>
 {
 public:
-    int translate(int x, int y);
+    int translate(float x, float y);
     int resize(float factor);
     int rotate(float degrees);
 
     Coordinates get_position();
-    std::array<int,2>  get_dimensions();
+    std::array<float,2>  get_dimensions();
 };
 
 bool operator==(const Element& e1, const Element& e2);
